fix: Moved fetch function inside utils and moved types and changed provider info
